Rename module: 'angular' -> 'angle'
